Use a Hugging Face pipeline or model_id to quickly predict on your inputs

If a model_id is provided, the Inference API will be used to make the prediction. If you wish to download a pipeline rather than running your predictions through the Inference API, download the model first with the hf_load_pipeline() function.

hf_predict(
  pipeline,
  inputs,
  parameters = NULL,
  use_gpu = FALSE,
  use_cache = FALSE,
  wait_for_model = FALSE,
  use_auth_token = NULL,
  ...
)

Arguments

pipeline

Either a downloaded pipeline from the Hugging Face Hub (using hf_load_pipeline()) or a model_id.

inputs

The data to predict on.

parameters

Model parameters distinct to the model being used.

use_gpu

API Only - Whether to use GPU for inference.

use_cache

API Only - Whether to use cached inference results for previously seen inputs.

wait_for_model

API Only - Whether to wait for the model to be ready instead of receiving a 503 error after a certain amount of time.

use_auth_token

API Only - The token to use as HTTP bearer authorization for the Inference API. Defaults to HUGGING_FACE_HUB_TOKEN environment variable.

...

sent to instantiation of model

Value

A Hugging Face model prediction.

See also