Search Models

Search Huggingface Models

hf_search_models(
  author = NULL,
  language = NULL,
  library = NULL,
  name = NULL,
  tags = NULL,
  task = NULL,
  dataset = NULL,
  search = NULL,
  sort = NULL,
  direction = NULL,
  limit = NULL,
  use_auth_token = NULL
)

Arguments

author

Filter by model author. Run hf_list_authors() for options.

language

Filter by the languages the model accommodates. Run hf_list_languages() for options.

library

Filter by the deep learning libraries which work with the model. Run hf_list_libraries() for options.

name

Filter by model names. Run hf_list_models() for options.

tags

Filter by model tags.

task

Filter by tasks the model can accomplish. Run hf_list_tasks() for options.

dataset

Filter by the datasets the model was trained on. hf_list_datasets()

search

A string that will be contained in the returned models

sort

The model parameter with which to sort the resulting models.

direction

Direction in which to sort. The value -1 sorts by descending order while all other values sort by ascending order.

limit

The limit on the number of models fetched. Leaving this option to NULL fetches all models.

use_auth_token

The token to use to access private repositories. Unnecessary if HUGGING_FACE_HUB_TOKEN environment variable is set. If True, will use the token generated when running transformers-cli login (stored in ~/.huggingface).

See also

Examples

hf_search_models(library = "pytorch", dataset = "mnli")
#> # A tibble: 27 × 12
#>    author         config downloads id    lastModified library_name likes modelId
#>    <chr>          <lgl>      <int> <chr> <chr>        <chr>        <int> <chr>  
#>  1 DeepPavlov     NA            52 Deep… 2021-11-15T… transformers     0 DeepPa…
#>  2 NbAiLab        NA           332 NbAi… 2021-11-17T… transformers     1 NbAiLa…
#>  3 digitalepidem… NA            56 digi… 2021-09-22T… transformers     0 digita…
#>  4 lighteternal   NA           171 ligh… 2021-10-27T… transformers     3 lighte…
#>  5 salesken       NA            45 sale… 2021-07-21T… transformers     2 salesk…
#>  6 valhalla       NA        644051 valh… 2021-06-14T… transformers    20 valhal…
#>  7 valhalla       NA         39712 valh… 2021-06-14T… transformers     7 valhal…
#>  8 valhalla       NA          7160 valh… 2021-06-14T… transformers     4 valhal…
#>  9 valhalla       NA          2771 valh… 2021-06-14T… transformers     2 valhal…
#> 10 vicgalle       NA          1522 vicg… 2021-03-04T… transformers    10 vicgal…
#> # … with 17 more rows, and 4 more variables: pipeline_tag <chr>, private <lgl>,
#> #   sha <chr>, tags <list>
hf_search_models(author = "facebook", name = "bart")
#> # A tibble: 18 × 12
#>    author      config downloads id       lastModified library_name likes modelId
#>    <chr>       <lgl>      <int> <chr>    <chr>        <chr>        <int> <chr>  
#>  1 facebook    NA       5392199 faceboo… 2022-11-16T… transformers    44 facebo…
#>  2 facebook    NA        919523 faceboo… 2022-11-23T… transformers   206 facebo…
#>  3 facebook    NA       1056307 faceboo… 2022-11-16T… transformers   257 facebo…
#>  4 facebook    NA         36121 faceboo… 2022-08-26T… transformers    14 facebo…
#>  5 facebook    NA        488808 faceboo… 2022-06-03T… transformers    22 facebo…
#>  6 textattack  NA             1 textatt… 2020-08-20T… transformers     0 textat…
#>  7 textattack  NA             2 textatt… 2020-08-20T… transformers     0 textat…
#>  8 textattack  NA            20 textatt… 2020-06-09T… transformers     0 textat…
#>  9 textattack  NA            12 textatt… 2020-06-09T… transformers     0 textat…
#> 10 textattack  NA            12 textatt… 2020-06-09T… transformers     0 textat…
#> 11 textattack  NA             5 textatt… 2020-06-09T… transformers     0 textat…
#> 12 textattack  NA            22 textatt… 2020-06-09T… transformers     0 textat…
#> 13 textattack  NA             4 textatt… 2020-06-09T… transformers     0 textat…
#> 14 textattack  NA             2 textatt… 2020-06-09T… transformers     0 textat…
#> 15 anuj55      NA             0 anuj55/… 2022-07-07T… NA               0 anuj55…
#> 16 minseok0809 NA             0 minseok… 2022-08-14T… NA               0 minseo…
#> 17 sunsvrv     NA            14 sunsvrv… 2022-10-10T… transformers     0 sunsvr…
#> 18 Yelyzaveta  NA             2 Yelyzav… 2022-10-28T… transformers     0 Yelyza…
#> # … with 4 more variables: pipeline_tag <chr>, private <lgl>, sha <chr>,
#> #   tags <list>