Load a dataset from the Hugging Face Hub!

Function has multiple uses - getting pre-made datasets for exploratory analysis, or to figure as means for evaluating your fine-tuned models.

hf_load_dataset(
  dataset,
  split = NULL,
  label_conversion = c("str2int", "int2str")
)

Arguments

dataset

The name of a Hugging Face dataset saved on the Hub. Use hf_list_models() to find a dataset.

split

Usually one of 'train' , 'test', 'validation' however, check the dataset's meta data at The Hub first.

label_conversion

Whether to add an additional column converting labels from str2int or int2str?

Value

A Hugging Face data set as a tibble

See also

Examples

(emotion_all_splits <- hf_load_dataset('emotion'))
#> $test
#> # A tibble: 2,000 × 3
#>    text                                                         label label_name
#>    <chr>                                                        <dbl> <chr>     
#>  1 im feeling rather rotten so im not very ambitious right now      0 sadness   
#>  2 im updating my blog because i feel shitty                        0 sadness   
#>  3 i never make her separate from me because i don t ever want…     0 sadness   
#>  4 i left with my bouquet of red and yellow tulips under my ar…     1 joy       
#>  5 i was feeling a little vain when i did this one                  0 sadness   
#>  6 i cant walk into a shop anywhere where i do not feel uncomf…     4 fear      
#>  7 i felt anger when at the end of a telephone call                 3 anger     
#>  8 i explain why i clung to a relationship with a boy who was …     1 joy       
#>  9 i like to have the same breathless feeling as a reader eage…     1 joy       
#> 10 i jest i feel grumpy tired and pre menstrual which i probab…     3 anger     
#> # … with 1,990 more rows
#> 
#> $train
#> # A tibble: 16,000 × 3
#>    text                                                         label label_name
#>    <chr>                                                        <dbl> <chr>     
#>  1 i didnt feel humiliated                                          0 sadness   
#>  2 i can go from feeling so hopeless to so damned hopeful just…     0 sadness   
#>  3 im grabbing a minute to post i feel greedy wrong                 3 anger     
#>  4 i am ever feeling nostalgic about the fireplace i will know…     2 love      
#>  5 i am feeling grouchy                                             3 anger     
#>  6 ive been feeling a little burdened lately wasnt sure why th…     0 sadness   
#>  7 ive been taking or milligrams or times recommended amount a…     5 surprise  
#>  8 i feel as confused about life as a teenager or as jaded as …     4 fear      
#>  9 i have been with petronas for years i feel that petronas ha…     1 joy       
#> 10 i feel romantic too                                              2 love      
#> # … with 15,990 more rows
#> 
#> $validation
#> # A tibble: 2,000 × 3
#>    text                                                         label label_name
#>    <chr>                                                        <dbl> <chr>     
#>  1 im feeling quite sad and sorry for myself but ill snap out …     0 sadness   
#>  2 i feel like i am still looking at a blank canvas blank piec…     0 sadness   
#>  3 i feel like a faithful servant                                   2 love      
#>  4 i am just feeling cranky and blue                                3 anger     
#>  5 i can have for a treat or if i am feeling festive                1 joy       
#>  6 i start to feel more appreciative of what god has done for …     1 joy       
#>  7 i am feeling more confident that we will be able to take ca…     1 joy       
#>  8 i feel incredibly lucky just to be able to talk to her           1 joy       
#>  9 i feel less keen about the army every day                        1 joy       
#> 10 i feel dirty and ashamed for saying that                         0 sadness   
#> # … with 1,990 more rows
#> 

(imdb_train <- hf_load_dataset('imdb', split = "train"))
#> # A tibble: 25,000 × 3
#>    text                                                         label label_name
#>    <chr>                                                        <dbl> <chr>     
#>  1 "I rented I AM CURIOUS-YELLOW from my video store because o…     0 neg       
#>  2 "\"I Am Curious: Yellow\" is a risible and pretentious stea…     0 neg       
#>  3 "If only to avoid making this type of film in the future. T…     0 neg       
#>  4 "This film was probably inspired by Godard's Masculin, fémi…     0 neg       
#>  5 "Oh, brother...after hearing about this ridiculous film for…     0 neg       
#>  6 "I would put this at the top of my list of films in the cat…     0 neg       
#>  7 "Whoever wrote the screenplay for this movie obviously neve…     0 neg       
#>  8 "When I first saw a glimpse of this movie, I quickly notice…     0 neg       
#>  9 "Who are these \"They\"- the actors? the filmmakers? Certai…     0 neg       
#> 10 "This is said to be a personal film for Peter Bogdonavitch.…     0 neg       
#> # … with 24,990 more rows