Search Models

Search Huggingface Models

hf_search_models(
  author = NULL,
  language = NULL,
  library = NULL,
  name = NULL,
  tags = NULL,
  task = NULL,
  dataset = NULL,
  search = NULL,
  sort = NULL,
  direction = NULL,
  limit = NULL,
  use_auth_token = NULL
)

Arguments

author

Filter by model author. Run hf_list_authors() for options.

language

Filter by the languages the model accommodates. Run hf_list_languages() for options.

library

Filter by the deep learning libraries which work with the model. Run hf_list_libraries() for options.

name

Filter by model names. Run hf_list_models() for options.

tags

Filter by model tags.

task

Filter by tasks the model can accomplish. Run hf_list_tasks() for options.

dataset

Filter by the datasets the model was trained on. hf_list_datasets()

search

A string that will be contained in the returned models

sort

The model parameter with which to sort the resulting models.

direction

Direction in which to sort. The value -1 sorts by descending order while all other values sort by ascending order.

limit

The limit on the number of models fetched. Leaving this option to NULL fetches all models.

use_auth_token

The token to use to access private repositories. Unnecessary if HUGGING_FACE_HUB_TOKEN environment variable is set. If True, will use the token generated when running transformers-cli login (stored in ~/.huggingface).

See also

Examples

hf_search_models(library = "pytorch", dataset = "mnli")
#> # A tibble: 26 × 12
#>    author         config downloads id    lastModified library_name likes modelId
#>    <chr>          <lgl>      <int> <chr> <chr>        <chr>        <int> <chr>  
#>  1 valhalla       NA        699370 valh… 2021-06-14T… transformers    16 valhal…
#>  2 valhalla       NA        147314 valh… 2021-06-14T… transformers     6 valhal…
#>  3 valhalla       NA         33284 valh… 2021-06-14T… transformers     4 valhal…
#>  4 vicgalle       NA         20175 vicg… 2021-03-04T… transformers    10 vicgal…
#>  5 valhalla       NA          5239 valh… 2021-06-14T… transformers    NA valhal…
#>  6 yoshitomo-mat… NA          2769 yosh… 2021-05-29T… transformers    NA yoshit…
#>  7 yoshitomo-mat… NA          1520 yosh… 2021-05-29T… transformers    NA yoshit…
#>  8 NbAiLab        NA           811 NbAi… 2021-11-17T… transformers     1 NbAiLa…
#>  9 digitalepidem… NA           328 digi… 2021-09-22T… transformers    NA digita…
#> 10 lighteternal   NA           119 ligh… 2021-10-27T… transformers     3 lighte…
#> # … with 16 more rows, and 4 more variables: pipeline_tag <chr>, private <lgl>,
#> #   sha <chr>, tags <list>
hf_search_models(author = "facebook", name = "bart")
#> # A tibble: 17 × 12
#>    author      config downloads id       lastModified library_name likes modelId
#>    <chr>       <lgl>      <int> <chr>    <chr>        <chr>        <int> <chr>  
#>  1 facebook    NA       6113715 faceboo… 2022-06-03T… transformers    28 facebo…
#>  2 facebook    NA       1012613 faceboo… 2022-08-08T… transformers   139 facebo…
#>  3 facebook    NA       1076337 faceboo… 2021-08-09T… transformers   200 facebo…
#>  4 facebook    NA         81667 faceboo… 2022-08-26T… transformers    11 facebo…
#>  5 facebook    NA        476769 faceboo… 2022-06-03T… transformers    16 facebo…
#>  6 textattack  NA             7 textatt… 2020-08-20T… transformers    NA textat…
#>  7 textattack  NA             3 textatt… 2020-08-20T… transformers    NA textat…
#>  8 textattack  NA            65 textatt… 2020-06-09T… transformers    NA textat…
#>  9 textattack  NA            24 textatt… 2020-06-09T… transformers    NA textat…
#> 10 textattack  NA             9 textatt… 2020-06-09T… transformers    NA textat…
#> 11 textattack  NA            55 textatt… 2020-06-09T… transformers    NA textat…
#> 12 textattack  NA             6 textatt… 2020-06-09T… transformers    NA textat…
#> 13 textattack  NA            31 textatt… 2020-06-09T… transformers    NA textat…
#> 14 textattack  NA            16 textatt… 2020-06-09T… transformers    NA textat…
#> 15 anuj55      NA             0 anuj55/… 2022-07-07T… NA              NA anuj55…
#> 16 minseok0809 NA             0 minseok… 2022-08-14T… NA              NA minseo…
#> 17 sunsvrv     NA            19 sunsvrv… 2022-10-10T… transformers    NA sunsvr…
#> # … with 4 more variables: pipeline_tag <chr>, private <lgl>, sha <chr>,
#> #   tags <list>