Search Models

Search Huggingface Models

hf_search_models(
  author = NULL,
  language = NULL,
  library = NULL,
  name = NULL,
  tags = NULL,
  task = NULL,
  dataset = NULL,
  search = NULL,
  sort = NULL,
  direction = NULL,
  limit = NULL,
  use_auth_token = NULL
)

Arguments

author

Filter by model author. Run hf_list_authors() for options.

language

Filter by the languages the model accommodates. Run hf_list_languages() for options.

library

Filter by the deep learning libraries which work with the model. Run hf_list_libraries() for options.

name

Filter by model names. Run hf_list_models() for options.

tags

Filter by model tags.

task

Filter by tasks the model can accomplish. Run hf_list_tasks() for options.

dataset

Filter by the datasets the model was trained on. hf_list_datasets()

search

A string that will be contained in the returned models

sort

The model parameter with which to sort the resulting models.

direction

Direction in which to sort. The value -1 sorts by descending order while all other values sort by ascending order.

limit

The limit on the number of models fetched. Leaving this option to NULL fetches all models.

use_auth_token

The token to use to access private repositories. Unnecessary if HUGGING_FACE_HUB_TOKEN environment variable is set. If True, will use the token generated when running transformers-cli login (stored in ~/.huggingface).

See also

Examples

hf_search_models(library = "pytorch", dataset = "mnli")
#> # A tibble: 24 × 12
#>    author     config downl…¹ id    lastM…² libra…³ likes modelId pipel…⁴ private
#>    <chr>      <lgl>    <int> <chr> <chr>   <chr>   <int> <chr>   <chr>   <lgl>  
#>  1 valhalla   NA      470874 valh… 2021-0… transf…    13 valhal… zero-s… FALSE  
#>  2 valhalla   NA      116071 valh… 2021-0… transf…     6 valhal… zero-s… FALSE  
#>  3 valhalla   NA       26398 valh… 2021-0… transf…     3 valhal… zero-s… FALSE  
#>  4 valhalla   NA        6426 valh… 2021-0… transf…    NA valhal… zero-s… FALSE  
#>  5 yoshitomo… NA        5618 yosh… 2021-0… transf…    NA yoshit… text-c… FALSE  
#>  6 vicgalle   NA        3755 vicg… 2021-0… transf…    10 vicgal… zero-s… FALSE  
#>  7 yoshitomo… NA        1269 yosh… 2021-0… transf…    NA yoshit… text-c… FALSE  
#>  8 NbAiLab    NA         712 NbAi… 2021-1… transf…     1 NbAiLa… zero-s… FALSE  
#>  9 digitalep… NA         243 digi… 2021-0… transf…    NA digita… zero-s… FALSE  
#> 10 lighteter… NA         209 ligh… 2021-1… transf…     3 lighte… text-c… FALSE  
#> # … with 14 more rows, 2 more variables: sha <chr>, tags <list>, and
#> #   abbreviated variable names ¹​downloads, ²​lastModified, ³​library_name,
#> #   ⁴​pipeline_tag
hf_search_models(author = "facebook", name = "bart")
#> # A tibble: 5 × 12
#>   author   config downloads id     lastM…¹ libra…² likes modelId pipel…³ private
#>   <chr>    <lgl>      <int> <chr>  <chr>   <chr>   <int> <chr>   <chr>   <lgl>  
#> 1 facebook NA       1425683 faceb… 2022-0… transf…    22 facebo… featur… FALSE  
#> 2 facebook NA       1066739 faceb… 2021-0… transf…   174 facebo… zero-s… FALSE  
#> 3 facebook NA        770129 faceb… 2022-0… transf…   124 facebo… summar… FALSE  
#> 4 facebook NA        327189 faceb… 2022-0… transf…    13 facebo… featur… FALSE  
#> 5 facebook NA        123482 faceb… 2022-0… transf…     8 facebo… summar… FALSE  
#> # … with 2 more variables: sha <chr>, tags <list>, and abbreviated variable
#> #   names ¹​lastModified, ²​library_name, ³​pipeline_tag